requirements.txt
numpy
scipy
pyaml
cadquery
cadquery-ocp
multimethod
typish
ezdxf
nptyping